fixed some tests now no http request is send in tests
