+++ src/app/app.component.spec.ts
@@ -12,7 +12,14 @@ import { AppComponent } from './app.component';
 import { BottomBarComponent } from './Components/bottom-bar/bottom-bar.component';
 import { AmyChatComponent } from './Components/bottom-bar/Components/amy-chat/amy-chat.component';
 
+import { CommandHandlerService } from './Components/bottom-bar/Services/command-handler.service';
+import { OptionsService } from './Services/options.service';
+import { ChatService } from './Components/bottom-bar/Components/amy-chat/Services/chat.service';
+
 describe('AppComponent', () => {
+    const optionsSpy = jasmine.createSpyObj('OptionsService', ['getLanguage']);
+    const commandHandlerSpy = jasmine.createSpyObj('CommandHandlerService', ['sendCommand']);
+    const chatSpy = jasmine.createSpyObj('ChatService', ['getMessages', 'addMessage']);
     beforeEach(async(() => {
         TestBed.configureTestingModule({
             declarations: [
@@ -26,6 +33,11 @@ describe('AppComponent', () => {
                 MatButtonModule,
                 MatToolbarModule,
                 RouterTestingModule
+            ],
+            providers: [
+                { provide: OptionsService, useValue: optionsSpy },
+                { provide: CommandHandlerService, useValue: commandHandlerSpy },
+                { provide: ChatService, useValue: chatSpy }
             ]
         }).compileComponents();
     }));
